requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.1
tqdm==4.66.1
transformers==4.36.2
# Using a pre-built version of FAISS that doesn't require compilation
faiss-cpu>=1.7.4
langchain==0.0.335
sentence-transformers==2.2.2
streamlit==1.28.0
python-dotenv==1.0.0
PyPDF2==3.0.1
numpy==1.26.1
scikit-learn==1.3.2
huggingface-hub==0.19.4
pydantic==2.5.2
torch>=2.2.0
lxml==4.9.3 